+++ changelog.md
@@ -4,6 +4,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+## <a name="1.8.0"/> 1.8.0 - 2025-04-20
+
+### Added
+
+- [#87](https://github.com/Azure/Microsoft.Extensions.Caching.Cosmos/pull/87) Support Cosmos DB SDK with System.Text.Json serialization.
+
 ## <a name="1.7.0"/> 1.7.0 - 2024-10-08
 
 ### Added